Move ReleaseTgg into its own source file
